backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.10.0
cryptography>=43.0.3
python-multipart>=0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
